feat: rework to make more use of the main run loop
* jobs cancelled whilst pending or prepared still do not run terminate() etc as a no-op, the ExecutorStates they produce allow them to skip over unnecessary steps.
* there is still some weirdness - for instance whether the CANCELLED_BY_USER is a final state & the returned ExecutorState vs get_state() for jobs terminated whilst prepared
* add a state diagram showing the effective high-level states of the combined local_executor/run system as a whole
* only copy logs into log_dir for cancelled jobs - I'm keep to land the other changes asap, this allows us to store the logs where admins can retrieve them & figure out how to surface them to users separately

+++ jobrunner/run.py
@@ -191,18 +191,6 @@ def handle_job(job, api, mode=None, paused=None):
     synchronous_transitions = getattr(api, "synchronous_transitions", [])
     is_synchronous = False
 
-    if job_definition.cancelled:
-        # cancelled is driven by user request, so is handled explicitly first
-        # regardless of executor state.
-        api.terminate(job_definition)
-
-        mark_job_as_failed(job, StatusCode.CANCELLED_BY_USER, "Cancelled by user")
-        # considered getting the usual loop to do this, but it wouldn't pick up
-        # jobs cancelled before starting (in UNKNOWN state)
-        api.finalize(job_definition)
-        api.cleanup(job_definition)
-        return
-
     # handle special modes before considering executor state, as they ignore it
     if paused:
         if job.state == State.PENDING:
@@ -242,6 +230,28 @@ def handle_job(job, api, mode=None, paused=None):
         return
     else:
         EXECUTOR_RETRIES.pop(job.id, None)
+
+    # cancelled is driven by user request, so is handled explicitly first
+    if job_definition.cancelled:
+        # if initial_status.state == ExecutorState.EXECUTED the job has already finished, so we
+        # don't need to do anything here
+        if initial_status.state == ExecutorState.EXECUTING:
+            api.terminate(job_definition)  # synchronous operation
+            new_status = api.get_status(job_definition)
+            new_statuscode, _default_message = STATE_MAP[new_status.state]
+            set_code(job, new_statuscode, "Cancelled whilst executing")
+            return
+        if initial_status.state == ExecutorState.PREPARED:
+            set_code(
+                job,
+                StatusCode.FINALIZED,
+                "Cancelled whilst prepared",
+            )
+            # Nb. no need to actually run finalize() in this case
+            return
+        if initial_status.state == ExecutorState.UNKNOWN:
+            mark_job_as_failed(job, StatusCode.CANCELLED_BY_USER, "Cancelled by user")
+            return
 
     # check if we've transitioned since we last checked and trace it.
     if initial_status.state in STATE_MAP:
@@ -338,10 +348,19 @@ def handle_job(job, api, mode=None, paused=None):
         new_status = api.finalize(job_definition)
 
     elif initial_status.state == ExecutorState.FINALIZED:
+        # Cancelled jobs that have had cleanup() should now be again set to cancelled here to ensure
+        # they finish in the FAILED state
+        if job_definition.cancelled:
+            mark_job_as_failed(job, StatusCode.CANCELLED_BY_USER, "Cancelled by user")
+            api.cleanup(job_definition)
+            return
+
         # final state - we have finished!
         results = api.get_results(job_definition)
+
         save_results(job, job_definition, results)
         obsolete = get_obsolete_files(job_definition, results.outputs)
+
         if obsolete:
             errors = api.delete_files(job_definition.workspace, Privacy.HIGH, obsolete)
             if errors:
@@ -355,6 +374,7 @@ def handle_job(job, api, mode=None, paused=None):
                 )
 
         api.cleanup(job_definition)
+
         # we are done here
         return
 
@@ -529,7 +549,9 @@ def mark_job_as_failed(job, code, message, error=None, **attrs):
     set_code(job, code, message, error=error, attrs=attrs)
 
 
-def set_code(job, code, message, error=None, results=None, timestamp_ns=None, **attrs):
+def set_code(
+    job, new_status_code, message, error=None, results=None, timestamp_ns=None, **attrs
+):
     """Set the granular status code state.
 
     We also trace this transition with OpenTelemetry traces.
@@ -547,8 +569,8 @@ def set_code(job, code, message, error=None, results=None, timestamp_ns=None, **
     else:
         timestamp_s = int(timestamp_ns / 1e9)
 
-    # if code has changed then trace it and update
-    if job.status_code != code:
+    # if status code has changed then trace it and update
+    if job.status_code != new_status_code:
 
         # handle timer measurement errors
         if job.status_code_updated_at > timestamp_ns:
@@ -560,24 +582,30 @@ def set_code(job, code, message, error=None, results=None, timestamp_ns=None, **
             log.warning(
                 f"negative state duration of {duration}, clamping to 1ms\n"
                 f"before: {job.status_code:<24} at {ns_timestamp_to_datetime(job.status_code_updated_at)}\n"
-                f"after : {code:<24} at {ns_timestamp_to_datetime(timestamp_ns)}\n"
+                f"after : {new_status_code:<24} at {ns_timestamp_to_datetime(timestamp_ns)}\n"
             )
             timestamp_ns = int(job.status_code_updated_at + 1e6)  # set duration to 1ms
             timestamp_s = int(timestamp_ns // 1e9)
 
         # update coarse state and timings for user
-        if code in [StatusCode.PREPARED, StatusCode.PREPARING]:
+        if new_status_code in [StatusCode.PREPARED, StatusCode.PREPARING]:
             # we've started running
             job.state = State.RUNNING
             job.started_at = timestamp_s
-        elif code.is_final_code:
+        elif new_status_code in [StatusCode.CANCELLED_BY_USER]:
+            # only set this cancelled status after any finalize/cleanup processes
+            job.state = State.FAILED
+        elif new_status_code.is_final_code:
             job.completed_at = timestamp_s
-            if code == StatusCode.SUCCEEDED:
+            if new_status_code == StatusCode.SUCCEEDED:
                 job.state = State.SUCCEEDED
             else:
                 job.state = State.FAILED
         # we sometimes reset the job back to pending
-        elif code in [StatusCode.WAITING_ON_REBOOT, StatusCode.WAITING_DB_MAINTENANCE]:
+        elif new_status_code in [
+            StatusCode.WAITING_ON_REBOOT,
+            StatusCode.WAITING_DB_MAINTENANCE,
+        ]:
             job.state = State.PENDING
             job.started_at = None
 
@@ -587,7 +615,7 @@ def set_code(job, code, message, error=None, results=None, timestamp_ns=None, **
         )
 
         # update db object
-        job.status_code = code
+        job.status_code = new_status_code
         job.status_message = message
         job.updated_at = timestamp_s
 
@@ -595,7 +623,7 @@ def set_code(job, code, message, error=None, results=None, timestamp_ns=None, **
         job.status_code_updated_at = timestamp_ns
         update_job(job)
 
-        if code.is_final_code:
+        if new_status_code.is_final_code:
             # transitioning to a final state, so just record that state
             tracing.record_final_state(
                 job,
